pyproject.toml
[project]
name = "text2sql-oracle"
version = "0.1.0"
description = "LLM-powered SQL generation for Oracle databases"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "oracledb>=2.0",
    "openai>=1.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "sentence-transformers>=3.0",
]
